add dropdown - add calculator screen

+++ composeApp/src/commonMain/kotlin/org/gabrieal/gymtracker/model/ProteinInput.kt
@@ -4,7 +4,7 @@ import org.gabrieal.gymtracker.util.enums.ActivityLevel
 import org.gabrieal.gymtracker.util.enums.FitnessGoal
 
 data class ProteinInput(
-    val weightKg: Double,
+    val weightKg: Int,
     val goal: FitnessGoal,
     val activityLevel: ActivityLevel
 )
